backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.25.0
scipy>=1.11.0
reportlab>=4.0.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
